+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Nothing for now.
 
 Compare with [last published version](https://github.com/meduzen/datetime-attribute/compare/1.3.3...main).
 
+### Fixed
+
+- Fix incorrect year for `datetime(date, 'week')` when the week started the previous year. For example, `2021-01-01` is a Friday and its week belongs to 2020 (as [per spec](./README.md#weeknumber). In that case, the output was `2021-53` instead of `2020-53`.
+
 ### Improved
 
 - Improve type definitions where `?` and `undefined` are redundant by removing `undefined`: